Move embedded C code to a header file
 #250
